Expose JSCode and add docs
